Move plot information to modal
